update allthemodium.json, removing the soul lava.
Removed the soul lava fluid from the config.
In this way, if you wish to add it later on after discussion, then it can be done so.
if I recall, some people were against it since it *does* apparently have some uses. So there ya go.

Only things remaining are just the biomes.